chore(pr630): tier-b cleanup — aria-label, dead postgres dep, dev-services bump
- user-menu.tsx: rename trigger aria-label "Settings" -> "Open user menu"
  so screen readers describe what the button actually does (it opens the
  user menu containing Settings + Log out, it does not jump to Settings).
- api/app/package.json: drop unused "postgres" catalog entry. No file
  under api/app/src/ imports postgres; the only repo-wide consumer is
  db/app/src/drivers/postgres.ts which keeps its own dep.
- pnpm-workspace.yaml: bump @lightfastai/dev-services catalog
  ^0.1.6 -> ^0.2.0 (intentional, bundled here to keep lockfile in sync).

Resolves CodeRabbit threads on user-menu.tsx aria-label and the
outside-diff comment on api/app/package.json.
