+++ api/app/package.json
@@ -52,7 +52,6 @@
     "@vendor/observability": "workspace:*",
     "drizzle-orm": "catalog:",
     "inngest": "catalog:",
-    "postgres": "catalog:",
     "superjson": "catalog:",
     "yaml": "^2.8.3",
     "zod": "catalog:"
