+++ pnpm-workspace.yaml
@@ -17,7 +17,7 @@ catalog:
   '@lightfastai/dev-cli': ^0.1.2
   '@lightfastai/dev-core': ^0.1.1
   '@lightfastai/dev-proxy': ^0.1.24
-  '@lightfastai/dev-services': ^0.1.6
+  '@lightfastai/dev-services': ^0.2.0
   '@neondatabase/serverless': ^1.1.0
   '@noble/ed25519': ^3.1.0
   '@noble/hashes': ^2.2.0
